Reorganize beaconState classes
